[Platform] Move `TokenUsage` tests to correct namespace
